Cleaned up store creation
