+++ src/_shared/modules/index.js
@@ -1,18 +1,6 @@
-import { createStore, applyMiddleware, compose } from 'redux';
-import { List, fromJS } from 'immutable';
-import createLogger from 'redux-logger';
-
 import { reducer } from './todo';
+const reducers = {
+  todo: reducer,
+};
 
-const logger = createLogger({
-  stateTransformer: object => fromJS(object).toJS(),
-  actionTransformer: object => fromJS(object).toJS(),
-  collapsed: true,
-  logErrors: false,
-});
-
-const createStoreWithMiddleware = compose(
-  applyMiddleware(logger)
-)(createStore);
-
-export default createStoreWithMiddleware(reducer, List());
+export default reducers;
